% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addTUNITS}
\alias{addTUNITS}
\title{addTUNITS}
\usage{
addTUNITS(stratum = NULL, towDist = 1.75, wingspread = 41)
}
\arguments{
\item{stratum}{the default is \code{NULL}. This is a dataframe of strata, each with an AREA field
in square nautical miles.}

\item{towDist}{the default is \code{1.75}.}

\item{wingspread}{the default is \code{41}.}
}
\value{
the original \code{stratum} object, but with an additional \code{TUNITS} field.
}
\description{
This function .
}
\note{
GSWARPOUT, GSCRUISELIST and GSSPEC can NOT be used to filter the other tables.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
