% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByTaxa.R
\name{aggregateByTaxa}
\alias{aggregateByTaxa}
\title{aggregateByTaxa}
\usage{
aggregateByTaxa(tblList = NULL, code = NULL, aphiaid = NULL, taxa = NULL, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}

\item{...}{other arguments passed to methods (e.g. 'debug' and 'quiet')}
}
\value{
#' @author  Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\description{
This function .
}
