% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\name{binnit}
\alias{binnit}
\title{binnit}
\usage{
binnit(data, bin_size = 1)
}
\arguments{
\item{data}{A data frame containing a numeric column named \code{FLEN}.}

\item{bin_size}{Numeric value specifying the width of each length bin. Default is \code{1}.}
}
\value{
The input data frame with an added column \code{FLEN_BIN} indicating the bin assignment for each fish length.
}
\description{
Assigns each FLEN to a length bin using the specified bin size. Bin values are calculated as the lower
edge of each bin, using the formula 1 + bin_size * floor(FLEN / bin_size). This approach is consistent with legacy
APL code, but bins do not necessarily start at the minimum observed length.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
importFrom dplyr mutate
}
