% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{calcTotalMean}
\alias{calcTotalMean}
\title{calcTotalMean}
\usage{
calcTotalMean(theDataByStrat = NULL, valueField = NULL, areaField = NULL)
}
\arguments{
\item{theDataByStrat}{the default is \code{NULL}. A data frame containing stratified data.}

\item{valueField}{the default is \code{NULL}. The name of the field containing values to be averaged.}

\item{areaField}{the default is \code{NULL}. The name of the field containing area values for weighting.}
}
\value{
A numeric value representing the area-weighted mean, rounded to 5 decimal places.
}
\description{
Calculate the area-weighted mean across strata.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
