% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{calcTotalSE_stratified}
\alias{calcTotalSE_stratified}
\title{calcTotalSE_stratified}
\usage{
calcTotalSE_stratified(
  theDataByStrat = NULL,
  valueField = NULL,
  areaField = NULL
)
}
\arguments{
\item{theDataByStrat}{the default is \code{NULL}. A data frame containing stratified data.}

\item{valueField}{the default is \code{NULL}. The name of the field containing standard error values.}

\item{areaField}{the default is \code{NULL}. The name of the field containing area values for weighting.}
}
\value{
A numeric value representing the stratified total standard error, rounded to 5 decimal places.
}
\description{
Calculate the total standard error for stratified data using area-weighted standard errors.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
