% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stranal.R
\name{calcTotalSE_unstratified}
\alias{calcTotalSE_unstratified}
\title{calcTotalSE_unstratified}
\usage{
calcTotalSE_unstratified(theDataByStrat = NULL, valueField = NULL)
}
\arguments{
\item{theDataByStrat}{the default is \code{NULL}. A data frame containing stratified data.}

\item{valueField}{the default is \code{NULL}. The name of the field containing standard error values to be summed.}
}
\value{
A numeric value representing the total standard error, rounded to 5 decimal places.
}
\description{
Calculate the total standard error for unstratified data by taking the square root of the sum of squared standard errors.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
