% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\name{correctForTowDist}
\alias{correctForTowDist}
\title{correctForTowDist}
\usage{
correctForTowDist(df, col, towDist = 1.75, distCol = "DIST")
}
\arguments{
\item{df}{the default is \code{NULL}. A data frame containing catch and distance data.}

\item{col}{the default is \code{NULL}. The name of the column containing values to be standardized.}

\item{towDist}{the default is \code{1.75}. The standard tow distance in nautical miles.}

\item{distCol}{the default is \code{"DIST"}. The name of the column containing actual tow distances.}
}
\value{
The input data frame with the specified column standardized and a new "_RAW" column containing original values.
}
\description{
Standardize catch values to a common tow distance. Creates a "_RAW" version of the original field and adjusts values proportionally to the standard tow distance.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
