% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDATRAS.R
\name{extractDATRAS}
\alias{extractDATRAS}
\title{Mar_DATRAS}
\usage{
extractDATRAS(
  years = NULL,
  survey = NULL,
  csv = T,
  cxn = NULL,
  data.dir = NULL,
  path = NULL,
  debug = F
)
}
\arguments{
\item{years}{default is \code{NULL}. This specifies the year(s) for which you'd like to generate
HH files. Single years are fine, as are vectors (e.g. \code{c(2011,1015)}, \code{2015:2019})}

\item{survey}{default is \code{NULL}. This specifies the survey(s) for which you'd like to generate
HH files.  Valid values are
\itemize{
\item \code{4X} - Type 1; Spring (i.e. months 1:4); 2008+; not strata 5Z* 
\item \code{GEORGES} - Type 1; Spring (i.e. months 1:4);  2008+; strata 5Z*
\item \code{SPRING} - Type 1; Spring (i.e. months 1:4); pre-2008; specific strata 
\item \code{4VSW}  - Type 1; Spring (i.e. months 1:4); 4VSW strata;  
\item \code{SUMMER} - Type 1; Summer (i.e. months 5:8); specific strata
\item \code{FALL} - Type 1; Fall (i.e. months 9:12)
}}

\item{csv}{default is \code{TRUE}.  If \code{TRUE}, csv files are generated for each HH code.  If
\code{FALSE}, the output exists only in the resultant list.}

\item{cxn}{A valid Oracle connection object. This parameter allows you to 
pass an existing connection, reducing the need to establish a new connection 
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{data.dir}{The default is \code{NULL}. This is the path to your Mar.datawrangling
rdata files}

\item{path}{the default is \code{NULL}. This is the path where you would like your output files created.}

\item{debug}{The default is \code{F}. Setting this to TRUE will limit the 
results to a single set for a single species.}
}
\value{
a list containing (named) objects - 1 for each generated HH file
}
\description{
This function generates ICES DATRAS-compatible files directly from the Maritimes
groundfish database.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{DATRAS}
