% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOBIS.R
\name{extractOBIS}
\alias{extractOBIS}
\title{extractOBIS}
\usage{
extractOBIS(survey = NULL, years = NULL, ...)
}
\arguments{
\item{survey}{the default is \code{NULL}. Valid values are "SPRING", "SUMMER", "FALL", and "4VSW".}

\item{years}{the default is \code{NULL}. Valid values range from 1970-present.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
a dataframe
}
\description{
This function generates an OBIS-compliant dataset, including "core", "occurrences"
and "emof" objects for a given survey.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca} \cr Heidi van Vliet
\email{Heidi.vanVliet@dfo-mpo.gc.ca}
}
