% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractOBIS_Events.R
\name{extractOBIS_Events}
\alias{extractOBIS_Events}
\title{extractOBIS_Events}
\usage{
extractOBIS_Events(survey = NULL, years = NULL, ...)
}
\arguments{
\item{survey}{the default is \code{NULL}. Valid values are "SPRING", "SUMMER", "FALL", and "4VSW".}

\item{years}{the default is \code{NULL}. Valid values range from 1970-present.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
a dataframe
}
\description{
This function generates an OBIS-compliant "events" object for a given survey for the 
specified years.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca} \cr Heidi van Vliet
\email{Heidi.vanVliet@dfo-mpo.gc.ca}
}
