% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\name{get_pesd_rvt_dir}
\alias{get_pesd_rvt_dir}
\title{get_pesd_rvt_dir}
\usage{
get_pesd_rvt_dir()
}
\value{
A character string representing the directory path "C:/DFO-MPO/PESDData/RVTransmogrifier". Creates the directory if it doesn't exist.
}
\description{
Get or create the directory path for RVTransmogrifier data storage.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
