% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\name{getSpInfo}
\alias{getSpInfo}
\title{getSpInfo}
\usage{
getSpInfo(taxa = NULL, tblList)
}
\arguments{
\item{taxa}{the default is \code{NULL}. Any value found in any of "SPEC", "KINGDOM", "PHYLUM", "CLASS", "ORDER", 
"FAMILY", or "GENUS" can be specified (e.g. \code{taxa=c("GADIDAE")})}

\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}
}
\value{
a dataframe of all species records where some taxonomic description matches the submitted value of "taxa".
}
\description{
This function returns any species that match the string sent to "taxa" for any of the recorded taxonomic
levels (e.g. "KINGDOM","PHYLUM","CLASS","ORDER","FAMILY","GENUS", "SPECIES" and common name.)
}
\examples{
\dontrun{
getSpInfo(taxa ="mich", tblList=data_2025)
    CODE APHIA_ID      COMM                          SPEC  KINGDOM     PHYLUM       CLASS
  604   126306 SNIPE EEL        NEMICHTHYS SCOLOPACEUS ANIMALIA   CHORDATA   TELEOSTEI
 2056   106182      <NA> ARCOSCALPELLUM MICHELOTTIANUM ANIMALIA ARTHROPODA THECOSTRACA
 
 ORDER        FAMILY          GENUS
 ANGUILLIFORMES NEMICHTHYIDAE     NEMICHTHYS
 SCALPELLOMORPHA  SCALPELLIDAE ARCOSCALPELLUM
}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
