% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvey.R
\name{getSurvey}
\alias{getSurvey}
\title{getSurvey}
\usage{
getSurvey(survey = NULL, years = NULL, ...)
}
\arguments{
\item{survey}{the default is \code{NULL}. This specifies which survey should be extracted.  Valid
values are "SPRING", "SUMMER", "FALL", and "4VSW".}

\item{years}{the default is \code{NULL}. This parameter allows you to generate datasets for one or
more specific years.  A value of NULL will result in products being generated for all years for
which data exists, and a vector of years will result in dataset that include the specified years.}

\item{...}{other arguments passed to methods (i.e. 'keep_nullsets', debug' and 'quiet')}
}
\value{
a list of dataframes which have been filtered to only include data related to the 
specified survey and years
}
\description{
A survey is a defined by a combination of a range of months, a selection of strata, 
and a tow 'type'. This function ensures that all those values are combined correctly to get the 
valid data for a survey.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
