% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW_sets.R
\name{NW_sets}
\alias{NW_sets}
\title{NW_sets}
\usage{
NW_sets(tblList = NULL, towDist = 1.75, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{towDist}{the default is \code{1.75}. This is ...}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function ...
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
