% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NW_strat.R
\name{NW_strat}
\alias{NW_strat}
\title{NW_strat}
\usage{
NW_strat(tblList = NULL, dfNWSets = NULL, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{dfNWSets}{this is the output from \code{NW_sets()}.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function ...
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
