% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwStrat.R
\name{nwStrat}
\alias{nwStrat}
\title{nwStrat}
\usage{
nwStrat(tblList = NULL, dfNWSets = NULL, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{dfNWSets}{this is the output from \code{NW_sets()}.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function calculates the stratified values of numbers and weights by strata
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
