% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRV.R
\name{plotRV}
\alias{plotRV}
\title{plotRV}
\usage{
plotRV(
  tblList = NULL,
  ptVar = "TOTWGT",
  strataVar = "MEAN_WGT",
  plotPts = TRUE,
  plotStrata = TRUE,
  code = NULL,
  aphiaid = NULL,
  title = "auto",
  hideTitle = FALSE,
  bathyRes = 4,
  showNullsets = TRUE,
  online = F,
  ...
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}

\item{ptVar}{the default is \code{"TOTWGT"}. Can also be \code{"TOTWGT"}.}

\item{strataVar}{the default is \code{"MEAN_WGT"}}

\item{plotPts}{the default is \code{TRUE}.}

\item{plotStrata}{the default is \code{TRUE}.}

\item{code}{the default is \code{NULL}. If data should be limited to a particular species, enter 
the species code here}

\item{aphiaid}{the default is \code{NULL}. If data should be limited to a particular aphiaid, 
enter the aphiaid here.}

\item{title}{the default is \code{"auto"}. By default, if the plot will be titled by the species 
available in the data. Changing this value will result in a custom title.}

\item{hideTitle}{the default is \code{FALSE}. If the automatic title (i.e common names of species 
present in the data) is not desired, set this to TRUE (or change the default value of \code{title})}

\item{bathyRes}{the default is \code{4}. This is the resolution of the bathymetric data, in 
minutes (sent directly to \code{marmap::getNOAA.bathy}).  If it is set to \code{NULL}, no 
bathymetric data will be plotted, and plots will be much faster.}

\item{showNullsets}{the default is \code{TRUE}.  If TRUE, sets where the specified species was not
caught will be plotted as '+'.  If FALSE, null sets will not be plotted.}

\item{online}{the default is \code{FALSE}.}

\item{...}{other arguments passed to methods (e.g. 'debug' and 'quiet')}
}
\value{
a ggplot object showing the positions of the data.  Null sets will be shown as + signs.
}
\description{
This function .
}
\note{
GSWARPOUT, GSCRUISELIST and GSSPEC can NOT be used to filter the other tables.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
