% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRVSummary.R
\name{plotRVSummary}
\alias{plotRVSummary}
\title{plotRVSummary}
\usage{
plotRVSummary(
  tblList = NULL,
  cxn = NULL,
  code = NULL,
  taxa = NULL,
  survey = NULL,
  strata = NULL,
  types = 1,
  species_name = NULL,
  stock = NULL,
  years = NULL,
  show_error = FALSE,
  use_new_areas = FALSE
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. A list of RV dataframes including GSINF, GSCAT, GSDET, GSMISSIONS, and GSSTRATUM.}

\item{cxn}{the default is \code{NULL}. A valid Oracle connection object. This parameter allows you to
pass an existing connection, reducing the need to establish a new connection
within the function. If provided, it takes precedence over the connection-
related parameters.}

\item{code}{the default is \code{NULL}. If data should be limited to a particular species, enter
the species code here}

\item{taxa}{the default is \code{NULL}. Any value found in any of "SPEC", "KINGDOM",
"PHYLUM", "CLASS", "ORDER", "FAMILY", or "GENUS" can be specified (e.g. \code{taxa=c("GADIDAE")})}

\item{survey}{the default is \code{NULL}. This specifies which survey should be extracted.  Valid
values are "SPRING", "SUMMER", "FALL", and "4VSW".  A value of NULL will result in products being
generated for all 4 different surveys.}

\item{strata}{the default is \code{NULL}. A vector of 1 or more strata can be passed to this function, and the
data will be filtered to only that/those strata (via GSSTRATUM$STRAT)}

\item{types}{the default is \code{1}.  A vector of 1 or more set "types" can be passed to this function, and the
data will be filtered to only that/those set types (via GSINF$TYPE)}

\item{species_name}{the default is \code{NULL}. This will be used in the plot title.}

\item{stock}{the default is \code{NULL}. A valid stock name from `stock_map`. If provided, it overrides
manually specified parameters like `code`, `months`, `strata`, etc.}

\item{years}{the default is \code{NULL}.}

\item{show_error}{the default is \code{FALSE}.  If confidence intervals should be shown on the plot, set this to TRUE.}

\item{use_new_areas}{the default is \code{FALSE}. This is a development parameter.  It won't work for most users.}
}
\value{
a list including a plot and the stratified biomass/abundance data behind the plot
}
\description{
plotRVSummary
}
