% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plot_gscat_sp.R
\name{qc_plot_gscat_sp}
\alias{qc_plot_gscat_sp}
\title{Multi-panel QA Plot for Vessel Conversion (GSCAT): Weight and Number, faceted by SPEC}
\usage{
qc_plot_gscat_sp(summary_df, spec, detailed_df = NULL)
}
\arguments{
\item{summary_df}{Output from qc_summary()$df_by_year}

\item{spec}{Value of SPEC to filter/plot}
}
\value{
Patchwork object, ready for print or save
}
\description{
Multi-panel QA Plot for Vessel Conversion (GSCAT): Weight and Number, faceted by SPEC
}
