% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{roundDD2Min}
\alias{roundDD2Min}
\title{roundDD2Min}
\usage{
roundDD2Min(x = NULL, how = "round", nearestMin = 1, digits = 4)
}
\arguments{
\item{x}{this is value that should be rounded.}

\item{how}{the default is \code{"round"}, but values of \code{"ceiling"} and \code{"floor"} are also 
acceptable. \code{"round"} just rounds the value, while the others forcibly round it up or down 
(respectively) to the nearest minute.}

\item{nearestMin}{the default is \code{1}, but values between 1 and 60 make sense.}

\item{digits}{the default is \code{4}.  This is how many decimal places the resultant data should be rounded to.}
}
\value{
the original value, but rounded.
}
\description{
This function can be used to round decimal degrees to the nearest geographic minute.  
It can be set to round up or down, and is useful for generating plots with boundaries that make sense.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
