% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvPlot.R
\name{rvPlot}
\alias{rvPlot}
\title{rvPlot}
\usage{
rvPlot(
  tblList = NULL,
  plotVar = "TOTNO",
  code = NULL,
  aphiaid = NULL,
  title = "auto",
  hideAutoTitle = FALSE,
  bathyRes = 4
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}

\item{plotVar}{the default is \code{"TOTNO"}. Can also be \code{"TOTWGT"}.}

\item{code}{the default is \code{NULL}. If data should be limited to a particular species, enter 
the species code here}

\item{aphiaid}{the default is \code{NULL}. If data should be limited to a particular aphiaid, 
enter the aphiaid here.}

\item{title}{the default is \code{"auto"}. By default, if the plot will be titled by the species 
available in the data. Changing this value will result in a custom title.}

\item{bathyRes}{the default is \code{4}. This is the resolution of the bathymetric data, in 
minutes (sent directly to \code{marmap::getNOAA.bathy})}

\item{hideTitle}{the default is \code{FALSE}. If the automatic title (i.e common names of species 
present in the data) is not desired, set this to TRUE (or change the default value of \code{title})}
}
\value{
a ggplot object showing the positions of the data.  Null sets will be shown as + signs.
}
\description{
This function .
}
\note{
GSWARPOUT, GSCRUISELIST and GSSPEC can NOT be used to filter the other tables.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
