% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-utils.R
\name{stock_list}
\alias{stock_list}
\title{Get a list of all stock names}
\usage{
stock_list()
}
\value{
A character vector of stock names.
}
\description{
This function returns all the stock names (keys) in the `stock_map` object.
}
\examples{
# Get all available stock names
stock_list()
}
