% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-utils.R
\name{stock_lookup}
\alias{stock_lookup}
\title{Retrieve stock information}
\usage{
stock_lookup(stock)
}
\arguments{
\item{stock}{A string representing the name of the stock. This must be one of the names in the `stock_map`.}
}
\value{
A list of stock details, including `code`, `months`, `strata`, `spec_name`, and `stock_name`.
}
\description{
This function retrieves detailed information about a specific stock from the `stock_map`.
}
\examples{
# Example: Retrieve information about the "GAMO_4Vn" stock
stock_lookup("GAMO_4Vn")
}
