% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_map}
\alias{stock_map}
\title{Predefined Stock Metadata Map}
\format{
A named list with elements for each stock:
\describe{
  \item{code}{Numeric code describing the stock.}
  \item{months}{Vector of relevant months.}
  \item{strata}{Strata for the stock, either numeric or character.}
  \item{spec_name}{Species name.}
  \item{stock_name}{Stock name, often combining geographic and biological information.}
}
}
\usage{
stock_map
}
\description{
A predefined list containing metadata about various fish stocks. The list includes
information such as `code`, `months`, `strata`, `spec_name`, and `stock_name`.
}
\examples{
# Access the stock map
stock_map
}
\keyword{datasets}
