% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stock-utils.R
\name{stock_validate}
\alias{stock_validate}
\title{Validate and retrieve stock information}
\usage{
stock_validate(stock)
}
\arguments{
\item{stock}{A string representing the stock name to validate.}
}
\value{
A list of stock details if the input is valid. If the input is invalid, an error is raised 
with suggestions for similar stock names, if available.
}
\description{
This function checks whether a given stock name is valid. If valid, it returns the stock's details. 
If invalid, it tries to find the closest matches (based on approximate matching) in the stock names, 
`spec_name`, or `stock_name`. If no close matches are found, an error is raised.
}
\examples{
# Validate a known stock name
stock_validate("GAMO_4Vn")

# Attempt to validate an incorrect stock name
# stock_validate("GAMO_4VX")  # Returns error: "Did you mean: GAMO_4Vn?"
}
