% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stranal.R
\name{stranal_detailed}
\alias{stranal_detailed}
\title{stranal_detailed}
\usage{
stranal_detailed(
  tblList,
  towDist = 1.75,
  by_sex = FALSE,
  conf_limits = 95,
  inc_limits = T,
  bin_size = 1
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. A list of RV dataframes including GSINF, GSCAT, and GSDET.}

\item{towDist}{the default is \code{1.75}. The standard tow distance in nautical miles used for standardization.}

\item{by_sex}{the default is \code{FALSE}. If TRUE, calculations are grouped by sex (FSEX) in addition to other grouping variables.}

\item{conf_limits}{the default is \code{95}.  This the value which will be used to calculate the upper and lower confidence intervals.}

\item{inc_limits}{the default is \code{TRUE}. This specifies if you would like to include the values for the upper and lower confidence intervals in your summary output.}

\item{bin_size}{the default is \code{1}.}
}
\value{
A list containing up to seven elements: stratified_bySet, stratified_byStrat, OVERALL_SUMMARY (from stratify_simple), plus length_set, length_strat (length-based summaries), age_set, and age_strat (age-based summaries). Length and age elements are only included if corresponding data exist.
}
\description{
Calculate detailed stratified estimates of biomass and abundance from RV survey data including length and age distributions. This function extends stratify_simple by incorporating detailed catch-at-length and catch-at-age data from GSDET, standardizing individual measurements, and generating stratified summaries by length and age classes.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
