% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{stratify}
\usage{
stratify(tblList = NULL, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{...}{other arguments passed to methods (i.e. 'towDist', 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function ...
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
