% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_calcLengths.R
\name{stratify_calcLengths}
\alias{stratify_calcLengths}
\title{stratify_calcLengths}
\usage{
stratify_calcLengths(
  tblList = NULL,
  dfNWSets = NULL,
  towDist = 1.75,
  bySex = F,
  ignoreLgrp = T
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}}

\item{dfNWSets}{this is the output from \code{NW_sets()}.}

\item{towDist}{the default is \code{1.75}. This is ...}

\item{bySex}{the default is \code{F}. This is ...}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function ...
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
