% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify_detailed}
\alias{stratify_detailed}
\title{stratify_detailed}
\usage{
stratify_detailed(tblList, towDist = 1.75, by_sex = FALSE)
}
\arguments{
\item{tblList}{the default is \code{NULL}. A list of RV dataframes including GSINF, GSCAT, and GSDET.}

\item{towDist}{the default is \code{1.75}. The standard tow distance in nautical miles used for standardization.}

\item{by_sex}{the default is \code{FALSE}. If TRUE, calculations are grouped by sex (FSEX) in addition to other grouping variables.}
}
\value{
A list containing up to seven elements: stratified_bySet, stratified_byStrat, OVERALL_SUMMARY (from stratify_simple), plus length_set, length_strat (length-based summaries), age_set, and age_strat (age-based summaries). Length and age elements are only included if corresponding data exist.
}
\description{
Calculate detailed stratified estimates of biomass and abundance from RV survey data including length and age distributions. This function extends stratify_simple by incorporating detailed catch-at-length and catch-at-age data from GSDET, standardizing individual measurements, and generating stratified summaries by length and age classes.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
