% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_nwSets.R
\name{stratify_nwSets}
\alias{stratify_nwSets}
\title{stratify_nwSets}
\usage{
stratify_nwSets(tblList = NULL, towDist = 1.75, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{towDist}{the default is \code{1.75}. This is the default tow length for the survey.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
...
}
\description{
This function calculates the stratified values of numbers and weights by set for 
every encountered taxa.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
