% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_prepare.R
\name{stratify_prepare}
\alias{stratify_prepare}
\title{stratify_prepare}
\usage{
stratify_prepare(tblList = NULL, ...)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes. Prior 
to running this function they should all have been filtered via \code{propagateChanges()}.}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}

\item{towDist}{the default is \code{1.75}. This is the default tow length for the survey.}
}
\value{
...
}
\description{
This function prepares a df for every encountered taxa that will be used for stratification.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
