% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-utils.R
\name{valPerSqKm}
\alias{valPerSqKm}
\title{valPerSqKm}
\usage{
valPerSqKm(theData = NULL, towDist_NM = 1.75, netWidth_ft = 41)
}
\arguments{
\item{theData}{the default is \code{NULL}. A numeric vector of catch values. NA values are converted to 0.}

\item{towDist_NM}{the default is \code{1.75}. The tow distance in nautical miles.}

\item{netWidth_ft}{the default is \code{41}. The net width (wingspread) in feet.}
}
\value{
A numeric vector of densities per square kilometer, rounded to 4 decimal places.
}
\description{
Convert catch values to density per square kilometer based on tow distance and net width.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
