% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRVData.R
\name{loadRVData}
\alias{loadRVData}
\title{loadRVData}
\usage{
loadRVData(...)
}
\arguments{
\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet')}
}
\value{
a list of the ~18 data frames from the RVSurveyData package
}
\description{
This function loads all of the tables from the RVSurveyData package into a single 
list.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
