% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRV.R
\name{plotRV}
\alias{plotRV}
\title{plotRV}
\usage{
plotRV(
  tblList = NULL,
  plotSets = "TOTWGT",
  plotNullSets = TRUE,
  plotCatchStrata = "MEAN_WGT",
  catchStrataData = NULL,
  plotStrata = TRUE,
  labelStrata = TRUE,
  plotBathy = "FILL",
  bathyIntervals = 200,
  plotNAFO = FALSE,
  labelNAFO = FALSE,
  ...
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}

\item{plotSets}{the default is \code{"TOTWGT"}, but can also be \code{"TOTWGT"}.  If set to \code{NULL},
no catch data points will be plotted.}

\item{plotNullSets}{the default is \code{TRUE}.  If TRUE, sets where the specified species was not
caught will be plotted as '+'.  If FALSE, null sets will not be plotted.}

\item{plotCatchStrata}{the default is \code{"MEAN_WGT"}, but can be any of \code{"TOTNO"}, \code{"TOTWGT"}, 
\code{"MEAN_WGT"}, \code{"MEAN_NO"}, \code{"BIOMASS"}, or \code{"ABUND"}. If set to \code{NULL},
no strata will be plotted by catch.}

\item{catchStrataData}{the default is \code{NULL}. This is a stratified object resulting from running 
\code{stratify}.  \code{stratify()} can generate multiple dfs within a list, but this parameter should
be pointed to a single dataframe.}

\item{plotStrata}{the default is \code{TRUE}. Should Maritimes strata be shown?}

\item{labelStrata}{the default is \code{TRUE}. If they are shown, should Maritimes strata be labelled?}

\item{plotBathy}{the default is \code{"FILL"}, but can be any of \code{"FILL"}, \code{"LINES"} or \code{NULL}.
FILL results in progressively darker blue areas as depth increases. 
LINES results in contour lines.
NULL does not show any bathymetric information.}

\item{bathyIntervals}{the default is \code{100} (m).  This is the difference in depth between contour lines.}

\item{plotNAFO}{the default is \code{TRUE}. Should NAFO areas be shown?}

\item{labelNAFO}{the default is \code{TRUE}. If they are shown, should NAFO areas be labelled?}

\item{...}{other arguments passed to methods (e.g.'taxa', 'code', 'aphiaid', debug' and 'quiet')}
}
\value{
a ggplot object showing the positions of the data.  Null sets will be shown as + signs.
}
\description{
This function can plot the list data resulting from \code{get_survey()} or from 
\code{propagateChanges()}.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
