% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagateChanges.R
\name{propagateChanges}
\alias{propagateChanges}
\title{propagateChanges}
\usage{
propagateChanges(
  tblList = NULL,
  keep_nullsets = T,
  survey = NULL,
  years = NULL,
  months = NULL,
  missions = NULL,
  strata = NULL,
  types = 1,
  areas = NULL,
  code = NULL,
  aphiaid = NULL,
  taxa = NULL,
  debug = F
)
}
\arguments{
\item{tblList}{the default is \code{NULL}. This is a list populated with all RV dataframes that
should have filtering applied to them.}

\item{keep_nullsets}{the default is \code{FALSE}. This is used to control whether or not the 
presence/absence of species changes the returned set locations. If \code{FALSE}, the only 
returned records will be directly linkable to the user-filtered records. If \code{TRUE}, set 
locations will still be returned, even if a particular species was not caught there.}

\item{years}{the default is \code{NULL}. A vector of 1 or more years can be passed to this function, and the data 
will be filtered to only that/those years  (via GSMISSIONS$YEAR).}

\item{months}{the default is \code{NULL}. A vector of 1 or more months can be passed to this function, and the data 
will be filtered to only that/those months (via GSINF$SDATE).}

\item{missions}{the default is \code{NULL}. A vector of 1 or more missions can be passed to this function, and the 
data will be filtered to only that/those missions (via GSMISSIONS$MISSION)}

\item{strata}{the default is \code{NULL}. A vector of 1 or more strata can be passed to this function, and the 
data will be filtered to only that/those strata (via GSSTRATUM$STRAT)}

\item{types}{the default is \code{NULL}. A vector of 1 or more set "types" can be passed to this function, and the 
data will be filtered to only that/those set types (via GSINF$TYPE)}

\item{areas}{the default is \code{NULL}. A vector of 1 or more areas can be passed to this function, and the 
data will be filtered to only that/those set types (via GSINF$AREA)}

\item{...}{other arguments passed to methods (i.e. 'debug' and 'quiet').}
}
\value{
a list of filtered versions of the dataframes passed as \code{tblList}. If the
filtering fails, a value of -1 will be returned. For example, if data is filtered for a year
where data was not collected, a strata that doesn't exist, or a species that was not observed
would all result in a value of -1 being returned.
}
\description{
This function is used to ensure that filtering of a single RV data object cascades
to all of the other known RV data objects. For example, if GSSPECIES is filtered to only
"haddock", and \code{propagateChanges} is run, then all of the various RV tables will be filtered to
the point where they relate directly to haddock. All remaining GSINF records would be sets that
caught haddock, all GSCAT records would be limited to haddock, all GSDET records would be limited
to haddock, etc. Filtering is not limited to species, but any value that exists in any field in
any of the tables used by this package (i.e."GSINF","GSCAT","GSMISSIONS","GSXTYPE","GSSTRATUM",
"GSWARPOUT","GSSPECIES_NEW","GSDET").
}
\note{
GSWARPOUT, GSCRUISELIST and GSSPEC can NOT be used to filter the other tables.
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
